requirements.txt
paramiko
public-ip
pyTelegramBotAPI
python-dotenv
redis
requests
SQLAlchemy
TelegramBotAPI
APScheduler
yookassa 